Rollback commons change until bugs are fixed (#654)
* Revert "Fixing commons update bugs (#653)"

This reverts commit ea0b0be960c7c2cb0baabfd2060b441a18e1bdc5.

* Revert "Update country_mappings.txt (#636)"

This reverts commit 3b21054b0d2bd53764ba1538f5e54ba240cbe6e2.

* Revert "Commons update (#652)"

This reverts commit 1fbd1667fe1c21b98128ed6f2e16e1c62d2ec0e0.

* Reapply "Update country_mappings.txt (#636)"

This reverts commit 018ffdbdfd7958c55054b6a3cb107fe21cb0eedf.

+++ EU4ToVic3/Source/V3World/EconomyManager/Demand/MarketTracker.cpp
@@ -156,7 +156,7 @@ void V3::MarketTracker::integrateBuilding(const Building& building,
 
 		lostSubsistence += marketJobs.createJobs(urbanResources.getJobs(),
 			 subsistenceResources.getJobs(),
-			 static_cast<int>(p * urbanFrac),
+			 p * urbanFrac,
 			 defaultRatio,
 			 addedWorkingPopFraction,
 			 {},
